requirements.txt
# Libraries for working with vector databases
# tiktoken
# datalake

flask
langchain
python-dotenv
openai
gunicorn
